BREAKING: Move NotificationType to its own class
